+++ HealthBarScripts/Boss/BossHealthBarController.cs
@@ -8,11 +8,27 @@ namespace SilkenImpact {
     public class BossHealthBarController : BaseHealthBarController<BossOwnerEvent, BossHealthBarOwner> {
 
         BossHealthBarContainer container;
-        protected string healthBarPrefabPath = "Assets/Addressables/Prefabs/HealthBarBossWithName.prefab";
+        protected string healthBarPrefabPath {
+            get {
+                switch (Configs.Instance.healthBarShape.Value) {
+                    case HealthBarShape.Rounded:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/RoundedBoss.prefab";
+                    case HealthBarShape.Diamond:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/DiamondBoss.prefab";
+                    default:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/RoundedBoss.prefab";
+                }
+            }
+        }
         protected string containerPrefabPath = "Assets/Addressables/Prefabs/Container.prefab";
         public override GameObject GetNewHealthBar => Plugin.InstantiateFromAssetsBundle(healthBarPrefabPath, "BossHealthBar");
 
         public override Canvas BarCanvas => ScreenSpaceCanvas.GetScreenSpaceCanvas;
+
+        void UpdateContainerWidth() {
+            float targetWidth = Configs.Instance.bossBarWidth.Value;
+            container.SetWidth(targetWidth);
+        }
 
         void prepareContainer() {
             var containerGO = Plugin.InstantiateFromAssetsBundle(containerPrefabPath, "BossHealthBarContainer");
@@ -28,8 +44,12 @@ namespace SilkenImpact {
 
         protected override void Awake() {
             base.Awake();
-
             prepareContainer();
+        }
+
+        protected override void MatchVisualsWithConfigs() {
+            base.MatchVisualsWithConfigs();
+            UpdateContainerWidth();
         }
 
 
@@ -37,16 +57,15 @@ namespace SilkenImpact {
             if (!guardExist(enemyGO)) return;
             base.OnEnemyShow(enemyGO);
 
-            var barGO = healthBarGoOf[enemyGO];
-            container.AddBar(barGO.GetComponent<HealthBar>());
+            var bar = healthBarOf[enemyGO];
+            container.AddBar(bar);
         }
 
         protected override void OnEnemySpawn(GameObject bossGO, float maxHp) {
             base.OnEnemySpawn(bossGO, maxHp);
 
-            var healthBarGO = healthBarGoOf[bossGO];
-            healthBarGO.transform.localScale = Vector3.one;
-            UIHealthBar uIHealthBar = healthBarGO.GetComponent<UIHealthBar>();
+            var healthBar = healthBarOf[bossGO];
+            UIHealthBar uIHealthBar = healthBar.GetComponent<UIHealthBar>();
             if (uIHealthBar) {
                 string locolisedName = HealthManagerPatch.LocalisedName(__instance: bossGO.GetComponent<HealthManager>());
                 PluginLogger.LogInfo($"BossHealthBarController: Localised name for bossGO {bossGO.name} is {locolisedName}");
@@ -60,14 +79,14 @@ namespace SilkenImpact {
             if (!guardExist(bossGO)) return;
             base.OnEnemyHide(bossGO);
 
-            var go = healthBarGoOf[bossGO];
-            container.RemoveBar(go.GetComponent<HealthBar>());
+            var bar = healthBarOf[bossGO];
+            container.RemoveBar(bar);
         }
 
         protected override void OnEnemyDie(GameObject bossGO) {
             if (!guardExist(bossGO)) return;
-            var go = healthBarGoOf[bossGO];
-            container.RemoveBar(go.GetComponent<HealthBar>());
+            var bar = healthBarOf[bossGO];
+            container.RemoveBar(bar);
 
             base.OnEnemyDie(bossGO);
         }