+++ HealthBarScripts/Mob/MobHealthBarController.cs
@@ -6,7 +6,18 @@ using UnityEngine.UI;
 
 namespace SilkenImpact {
     public class MobHealthBarController : BaseHealthBarController<MobOwnerEvent, MobHealthBarOwner> {
-        private string prefabPath = "Assets/Addressables/Prefabs/HealthBarSmall.prefab";
+        private string prefabPath {
+            get {
+                switch (Configs.Instance.healthBarShape.Value) {
+                    case HealthBarShape.Rounded:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/Rounded.prefab";
+                    case HealthBarShape.Diamond:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/Diamond.prefab";
+                    default:
+                        return "Assets/Addressables/Prefabs/Health Bars/Masked UI Health Bars/Rounded.prefab";
+                }
+            }
+        }
         public override GameObject GetNewHealthBar => Plugin.InstantiateFromAssetsBundle(prefabPath, "MobHealthBar");
 
         public override Canvas BarCanvas => WorldSpaceCanvas.GetWorldSpaceCanvas;
@@ -17,21 +28,25 @@ namespace SilkenImpact {
 
         protected override void OnEnemySpawn(GameObject enemyGO, float maxHp) {
             base.OnEnemySpawn(enemyGO, maxHp);
-            var healthBarGO = healthBarGoOf[enemyGO];
-            var spriteTracker = healthBarGO.GetComponent<SpriteTracker>();
+            var healthBarGO = healthBarOf[enemyGO].gameObject;
+            if (!healthBarGO.TryGetComponent<SpriteTracker>(out var spriteTracker)) {
+                spriteTracker = healthBarGO.AddComponent<SpriteTracker>();
+            }
             spriteTracker.SetTarget(enemyGO);
             spriteTracker.zOffset = Configs.Instance.spriteTrackerZOffset.Value;
         }
 
 #if DEBUG
         void Update() {
-            foreach (var kvp in healthBarGoOf) {
+            foreach (var kvp in healthBarOf) {
                 var hm = kvp.Key.GetComponent<HealthManager>();
-                var barGO = kvp.Value;
-                var bar = barGO.GetComponent<HealthBar>();
-                if (hm == null || barGO == null) continue;
+                var bar = kvp.Value;
+                var barGO = bar.gameObject;
+                if (hm == null || barGO == null)
+                    continue;
                 barGO.name = hm.name + "_HealthBar";
-                if (bar) bar.name = hm.name + "_HealthBar_Component";
+                if (bar)
+                    bar.name = hm.name + "_HealthBar_Component";
             }
         }