+++ Src/BossHealthBarContainer.cs
@@ -6,6 +6,13 @@ using UnityEngine;
 using UnityEngine.Assertions.Must;
 
 public class BossHealthBarContainer : MonoBehaviour {
+    /*
+    Important Rect Transform Settings of a BossHealthBar:
+    Anchor Min: (0, 0.5)
+    Anchor Max: (0, 0.5)
+    Pivot: (0.5, 0.5)
+    Make sure these are correctly set in the prefab!
+    */
     [SerializeField] private List<HealthBar> bars = new();
     [SerializeField] private RectTransform rect;
     public float interval = 0.3f;
@@ -28,6 +35,12 @@ public class BossHealthBarContainer : MonoBehaviour {
         OnUpdate();
     }
 #endif
+
+    public void SetWidth(float width) {
+        var r = GetComponent<RectTransform>();
+        r.sizeDelta = new Vector2(width, r.sizeDelta.y);
+        OnUpdate();
+    }
 
     public void AddBar(HealthBar bar) {
         if (bars.Find(b => b == bar)) return;